Move MapRequestPayload functional tests to separate controller
