+++ tests/Functional/SymfonyFunctionalTest.php
@@ -13,7 +13,6 @@ namespace Nelmio\ApiDocBundle\Tests\Functional;
 
 use Symfony\Component\HttpKernel\Attribute\MapQueryParameter;
 use Symfony\Component\HttpKernel\Attribute\MapQueryString;
-use Symfony\Component\HttpKernel\Attribute\MapRequestPayload;
 
 class SymfonyFunctionalTest extends WebTestCase
 {
@@ -406,141 +405,6 @@ class SymfonyFunctionalTest extends WebTestCase
         ], json_decode($this->getOperation('/api/article_map_query_parameter_overwrite_parameters', 'get')->toJson(), true));
     }
 
-    public function testMapRequestPayload(): void
-    {
-        if (!class_exists(MapRequestPayload::class)) {
-            self::markTestSkipped('Symfony 6.3 MapRequestPayload attribute not found');
-        }
-
-        self::assertEquals([
-            'operationId' => 'post_api_nelmio_apidoc_tests_functional_api_createarticlefrommaprequestpayload',
-            'responses' => [
-                '200' => [
-                    'description' => '',
-                ],
-            ],
-            'requestBody' => [
-                'content' => [
-                    'application/json' => [
-                        'schema' => [
-                            '$ref' => '#/components/schemas/Article81',
-                        ],
-                    ],
-                ],
-                'required' => true,
-            ],
-        ], json_decode($this->getOperation('/api/article_map_request_payload', 'post')->toJson(), true));
-    }
-
-    public function testMapRequestPayloadNullable(): void
-    {
-        if (!class_exists(MapRequestPayload::class)) {
-            self::markTestSkipped('Symfony 6.3 MapRequestPayload attribute not found');
-        }
-
-        self::assertEquals([
-            'operationId' => 'post_api_nelmio_apidoc_tests_functional_api_createarticlefrommaprequestpayloadnullable',
-            'responses' => [
-                '200' => [
-                    'description' => '',
-                ],
-            ],
-            'requestBody' => [
-                'content' => [
-                    'application/json' => [
-                        'schema' => [
-                            'nullable' => true,
-                            'oneOf' => [
-                                ['$ref' => '#/components/schemas/Article81'],
-                            ],
-                        ],
-                    ],
-                ],
-                'required' => false,
-            ],
-        ], json_decode($this->getOperation('/api/article_map_request_payload_nullable', 'post')->toJson(), true));
-    }
-
-    public function testMapRequestPayloadOverwriteRequestBody(): void
-    {
-        if (!class_exists(MapRequestPayload::class)) {
-            self::markTestSkipped('Symfony 6.3 MapRequestPayload attribute not found');
-        }
-
-        self::assertEquals([
-            'operationId' => 'post_api_nelmio_apidoc_tests_functional_api_createarticlefrommaprequestpayloadoverwrite',
-            'responses' => [
-                '200' => [
-                    'description' => '',
-                ],
-            ],
-            'requestBody' => [
-                'content' => [
-                    'application/json' => [
-                        'schema' => [
-                            '$ref' => '#/components/schemas/EntityWithNullableSchemaSet',
-                        ],
-                    ],
-                ],
-                'required' => true,
-                'description' => 'Request body description',
-            ],
-        ], json_decode($this->getOperation('/api/article_map_request_payload_overwrite', 'post')->toJson(), true));
-    }
-
-    public function testMapRequestPayloadHandlesAlreadySetContent(): void
-    {
-        if (!class_exists(MapRequestPayload::class)) {
-            self::markTestSkipped('Symfony 6.3 MapRequestPayload attribute not found');
-        }
-
-        self::assertEquals([
-            'operationId' => 'post_api_nelmio_apidoc_tests_functional_api_createarticlefrommaprequestpayloadhandlesalreadysetcontent',
-            'responses' => [
-                '200' => [
-                    'description' => '',
-                ],
-            ],
-            'requestBody' => [
-                'content' => [
-                    'application/json' => [
-                        'schema' => [
-                            '$ref' => '#/components/schemas/Article81',
-                        ],
-                    ],
-                ],
-                'required' => true,
-                'description' => 'Request body description',
-            ],
-        ], json_decode($this->getOperation('/api/article_map_request_payload_handles_already_set_content', 'post')->toJson(), true));
-    }
-
-    public function testMapRequestPayloadPassesValidationGroups(): void
-    {
-        if (!class_exists(MapRequestPayload::class)) {
-            self::markTestSkipped('Symfony 6.3 MapRequestPayload attribute not found');
-        }
-
-        self::assertEquals([
-            'operationId' => 'post_api_nelmio_apidoc_tests_functional_api_createarticlefrommaprequestpayloadpassedvalidationgroups',
-            'responses' => [
-                '200' => [
-                    'description' => '',
-                ],
-            ],
-            'requestBody' => [
-                'content' => [
-                    'application/json' => [
-                        'schema' => [
-                            '$ref' => '#/components/schemas/SymfonyConstraintsTestGroup',
-                        ],
-                    ],
-                ],
-                'required' => true,
-            ],
-        ], json_decode($this->getOperation('/api/article_map_request_payload_validation_groups', 'post')->toJson(), true));
-    }
-
     public function testMapQueryStringPassesValidationGroups(): void
     {
         if (!class_exists(MapQueryString::class)) {
@@ -829,72 +693,5 @@ class SymfonyFunctionalTest extends WebTestCase
                 ],
             ],
         ], json_decode($this->getOperation('/api/article_map_query_parameter_validate_filters', 'get')->toJson(), true)['parameters']);
-    }
-
-    public function testMapRequestPayloadArray(): void
-    {
-        if (!class_exists(MapRequestPayload::class)) {
-            self::markTestSkipped('Symfony 6.3 MapRequestPayload attribute not found');
-        }
-
-        if (!property_exists(MapRequestPayload::class, 'type')) {
-            self::markTestSkipped('Symfony 7.1 MapRequestPayload attribute type property not found');
-        }
-
-        self::assertEquals([
-            'operationId' => 'post_api_nelmio_apidoc_tests_functional_api_createarticlefrommaprequestpayloadarray',
-            'responses' => [
-                '200' => [
-                    'description' => '',
-                ],
-            ],
-            'requestBody' => [
-                'content' => [
-                    'application/json' => [
-                        'schema' => [
-                            'type' => 'array',
-                            'items' => [
-                                '$ref' => '#/components/schemas/Article81',
-                            ],
-                        ],
-                    ],
-                ],
-                'required' => true,
-            ],
-        ], json_decode($this->getOperation('/api/article_map_request_payload_array', 'post')->toJson(), true));
-    }
-
-    public function testMapRequestPayloadNullableArray(): void
-    {
-        if (!class_exists(MapRequestPayload::class)) {
-            self::markTestSkipped('Symfony 6.3 MapRequestPayload attribute not found');
-        }
-
-        if (!property_exists(MapRequestPayload::class, 'type')) {
-            self::markTestSkipped('Symfony 7.1 MapRequestPayload attribute type property not found');
-        }
-
-        self::assertEquals([
-            'operationId' => 'post_api_nelmio_apidoc_tests_functional_api_createarticlefrommaprequestpayloadnullablearray',
-            'responses' => [
-                '200' => [
-                    'description' => '',
-                ],
-            ],
-            'requestBody' => [
-                'content' => [
-                    'application/json' => [
-                        'schema' => [
-                            'type' => 'array',
-                            'items' => [
-                                '$ref' => '#/components/schemas/Article81',
-                            ],
-                            'nullable' => true,
-                        ],
-                    ],
-                ],
-                'required' => false,
-            ],
-        ], json_decode($this->getOperation('/api/article_map_request_payload_nullable_array', 'post')->toJson(), true));
     }
 }
